refactor(table): extract operation conflict checks into its own module
PartialEq for Operation is 836 lines: one arm per pair of operation variants,
hand-written because several operations carry Vec fields whose order is not
meaningful. It sat between the Operation definition and the logic that consumes
it, along with the four config-key helpers the commit retry path uses to decide
whether a concurrent operation touched metadata it depends on.

Grouping them puts every "do these two operations collide" question in one file.
